+++ chronosphere/intschema/log_ingest_config.go
@@ -14,7 +14,8 @@ import (
 var _ tfid.ID // Always use tfid for simplified import generation.
 
 type LogIngestConfig struct {
-	Parser []LogIngestConfigParser `intschema:"parser,optional"`
+	FieldParser     []LogIngestConfigFieldParser     `intschema:"field_parser,optional"`
+	PlaintextParser []LogIngestConfigPlaintextParser `intschema:"plaintext_parser,optional"`
 
 	// Internal identifier used in the .state file, i.e. ResourceData.Id().
 	// Cannot be set, else ToResourceData will panic.
@@ -53,7 +54,16 @@ func (o *LogIngestConfig) Ref() tfid.ID {
 	}.AsID()
 }
 
-type LogIngestConfigParser struct {
-	Name  string `intschema:"name"`
-	Regex string `intschema:"regex"`
+type LogIngestConfigPlaintextParser struct {
+	Name         string     `intschema:"name"`
+	DropOriginal bool       `intschema:"drop_original,optional"`
+	Mode         string     `intschema:"mode,optional"`
+	Parser       *LogParser `intschema:"parser,optional,list_encoded_object"`
+}
+
+type LogIngestConfigFieldParser struct {
+	Destination *LogFieldPath `intschema:"destination,optional,list_encoded_object"`
+	Mode        string        `intschema:"mode,optional"`
+	Parser      *LogParser    `intschema:"parser,optional,list_encoded_object"`
+	Source      *LogFieldPath `intschema:"source,optional,list_encoded_object"`
 }